refact: RacingCars의 네이밍 규칙에 대한 테스트를 RacingCarNameTest로 이관
