+++ src/test/java/racingcar/model/RacingCarsTest.java
@@ -18,14 +18,6 @@ class RacingCarsTest {
     }
 
     @Test
-    @DisplayName("RacingCars 객체를 만들 때 5자 이상의 이름을 입력하면 예외가 발생한다.")
-    void create_racing_cars_negative() throws Exception {
-        assertThatThrownBy(() -> RacingCars.create("qwertyasdfg"))
-                .isInstanceOf(RuntimeException.class)
-                .hasMessage("자동차 이름은 5자를 초과할 수 없습니다.");
-    }
-
-    @Test
     @DisplayName("RacingCars 객체를 만들 때 이름을 입력하지 않으면 예외가 발생한다.")
     void create_racing_cars_negative_blank() throws Exception {
         assertThatThrownBy(() -> RacingCars.create(""))
